Time: 3 ms (45.33%), Space: 8.5 MB (91.95%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         if(x<0) return 0;
-        long long int rev=0;
+        long int rev=0;
         int temp =x;
         while(temp!=0){
             int rem = temp %10;
